rework Tui->connection communication path
Previously each server had a unique Box<Conn> object that it used to communicate with the backend service.This is not a viable design for Discord, where there is a single connection to the Discord service, and multiple servers send and are sent messages via that single connection.

Now, a service sends the Tui a ServerConnected event with a mpsc::SyncSender that it should be sent events on. For a Discord or other such service, these could all be clones, thus enabling centralized handling of the TuiEvents that will be sent on it to send messages and such. Additionally, all the TuiEvents now indicate the server they originated from, so that a backend that handles multiple servers can act appropriately.
As an aside, I've factored out the autocompletion mechanics into their own object for now because it doesn't make sense to do non-blocking autocompletion by sending messages over a channel.

This change does *not* enable the use of an immutable internal ID to identify channels or servers which was another possible solution to the Discord architecture. That may still be useful in the future, but this architecture does produce some consistency in the way that a backend communicates with the Tui- the previous setup basically required that the backend not block on having any of its methods called which is an odd requirement.